+++ doc/release-notes.md
@@ -5,3 +5,5 @@ Bitcoin ABC version 0.26.3 is now available from:
   <https://download.bitcoinabc.org/0.26.3/>
 
 This release includes the following features and fixes:
+- A new `total_fee` field showing the total fees for all transactions in the
+  mempool has been added to the `getmempoolinfo` RPC.